+++ lib/rrj/tools/bin/init.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require 'rrj/rails' if defined?(::Rails::Engine)
+require 'rrj/rails'
 
-Log.info 'Listen Public queue to Janus'
 loop {}